requirements.txt
opencv-python
matplotlib
pillow
ultralytics
tensorflow[and-cuda]
pandas
numpy==1.26.4
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0